Move preserve game to utils Delete preserve file
