+++ data/albums.json
@@ -1,1 +1,1 @@
-[]
+[{"genre_id":"d6f2cd73-6552-417e-821c-6e09998a06f6","author_id":"43c37e15-ee26-4aaa-b4af-68f47d82c0ff","label_id":"06673eca-f2fa-44ef-bc0c-9748c8371609","publish_date":"2022-11-11","on_spotify":false}]
