+++ data/authors.json
@@ -1,1 +1,1 @@
-[["Jonh","Smith","441ce3e5-f8e9-4623-88d8-71e0d0307935"],["Mario","Bros","f4576079-7be6-4029-8c69-018e2fd6a42f"]]
+[["Michael","Jackson","fefbe279-6709-4628-ae6c-454afddc2100"],["Franco","Villa","43c37e15-ee26-4aaa-b4af-68f47d82c0ff"],["Jonh","Smith","441ce3e5-f8e9-4623-88d8-71e0d0307935"],["Mario","Bros","f4576079-7be6-4029-8c69-018e2fd6a42f"],["Little","Jimmy","827661ff-b253-454d-a096-c51943b44f49"],["Pac","Man","8c806b03-e7d6-44be-9e0c-d2489856d948"]]
