+++ data/books.json
@@ -1,1 +1,1 @@
-[]
+[{"id":"89c0097f-59ab-4f00-829a-b11e8f71e949","genre_id":"28e88a65-1bb1-435b-ad66-8809eaa4da3d","author_id":"fefbe279-6709-4628-ae6c-454afddc2100","label_id":"6ec50c12-2536-4a22-a0d5-b475f901c5cd","publish_date":"2022-01-01","publisher":"frut","cover_state":"bad"},{"id":"601d5ebb-86c8-4b41-8afb-4d871da06918","genre_id":"f2a017d4-7eb2-491e-872a-f9cae47dd980","author_id":"827661ff-b253-454d-a096-c51943b44f49","label_id":"9ad12c2a-bc49-47f0-8b08-2d2ad58e70ed","publish_date":"2022-01-01","publisher":"2020-01-01","cover_state":"bad"}]
