+++ data/genres.json
@@ -1,1 +1,1 @@
-[{"id":"64d63c57-310d-4477-862c-b2496a7baf4f","name":"terror"},{"id":"73fff21e-2bbf-4bc3-8537-803c947c2265","name":"none"}]
+[{"id":"d6f2cd73-6552-417e-821c-6e09998a06f6","name":"blues"},{"id":"64d63c57-310d-4477-862c-b2496a7baf4f","name":"terror"},{"id":"73fff21e-2bbf-4bc3-8537-803c947c2265","name":"none"},{"id":"f2a017d4-7eb2-491e-872a-f9cae47dd980","name":"fun"},{"id":"28e88a65-1bb1-435b-ad66-8809eaa4da3d","name":"rock"},{"id":"851a71bc-e39e-49e6-a125-7d580bd93cee","name":"arcade"}]
