+++ data/labels.json
@@ -1,1 +1,1 @@
-[{"title":"Shadow","color":"black","items":[],"id":"c6224197-7a62-4d67-bfdb-466908a03033"},{"title":"Mono","color":"red","items":[],"id":"bd3fad3b-9632-4887-8f35-6a4504839a3f"}]
+[{"title":"Bad","color":"red","items":[],"id":"6ec50c12-2536-4a22-a0d5-b475f901c5cd"},{"title":"Mine","color":"green","items":[],"id":"06673eca-f2fa-44ef-bc0c-9748c8371609"},{"title":"Shadow","color":"black","items":[],"id":"c6224197-7a62-4d67-bfdb-466908a03033"},{"title":"Mono","color":"red","items":[],"id":"bd3fad3b-9632-4887-8f35-6a4504839a3f"},{"title":"My toast","color":"blue","items":[],"id":"9ad12c2a-bc49-47f0-8b08-2d2ad58e70ed"},{"title":"Pacman","color":"yellow","items":[],"id":"134e5310-dd72-4ec9-939f-001a4d1ac7df"}]
